fix: use `instanceof` elytra item check for compat with modded elytras
Using `.getItem() instanceof ElytraItem` instead of `.is(Items.ELYTRA)`
allows the swapping to function with modded elytra items that extend
the vanilla `ElytraItem` class.

I also changed the Minecraft version dependency for the Fabric side to
`1.20.x` so that the mod can still be used with earlier 1.20 versions. I
only tested that it's working on 1.20.2 and I didn't test Forge at all.